Implement Compiler and VM, dummy world exploration
Implements a compiler and runtime for logic rules. It's all in pretty rough shape, and needs some cleaning up and relocating of pieces but its able to take the logic files, compile them and then execute all of them successfully -- for the given, hardcoded settings. 

Also implemented is a dummy world exploration which marries these compiled rules, a VM instance, a world graph, and a dummied world/VM state. This allows all nodes to be reached and ergo all rule edges to executed (again for the hardcoded settings).